HostNetwork: split the veth creation and vrf enslaving logic
We want to be able to create veths without a VRF. So we split and clean
the logic. The enslavement in l3vni was performed also in the SetupL3VNI
function. The enslavement in SetupL2VNI was not necessary as the veth
leg is enslaved to the bridge afterwards.

Additionally, we propagate the names from outside instead of calculating
them from the vni. This allows to create veths for multiple purpouses.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>